requirements.txt
selenium
pytest
pytest-fixtures
pytest-html
pytest-ordering
pytest-rerunfailures
xlrd
ddt
cx_Oracle
pytest-reporter-html1
pytest-xdist
allure-pytest
webdriver-manager